[Bug #11691] Several key values missing from text overlay for Hologic DBTs * Fixed.  * Apparently the compression force item was forcing an integer format, which    meant values with decimal points did not render at all.
--HG--
branch : Rigel
extra : rebase_source : 0e57f93b3d1faf1f4582f39042136bba730623cd

+++ ImageViewer/AnnotationProviders/Dicom/DXImageAnnotationItemProvider.cs
@@ -117,7 +117,7 @@ namespace ClearCanvas.ImageViewer.AnnotationProviders.Dicom
 							{
 								var acquisitions = _frameContext.GetData<IXRay3DAcquisitionSequenceItem[]>(f, _keyAcquisitionSequence);
 								var dataset = acquisitions != null ? (IEnumerable<IDicomAttributeProvider>) acquisitions.Select(a => a.DicomSequenceItem) : new IDicomAttributeProvider[] {f};
-								return FormatMultiValues(dataset, FormatInt32, DicomTags.CompressionForce, SR.FormatNewtons);
+								return FormatMultiValues(dataset, FormatFloat64, DicomTags.CompressionForce, SR.FormatNewtons);
 							},
 						DicomDataFormatHelper.RawStringFormat
 						)
